chore(origin-ui-organization): adjust modals flow to match app loading state
