+++ packages/ui/libs/organization/data/src/handlers/receivedInvitationsHandlers.ts
@@ -21,12 +21,7 @@ export const useReceivedInvitationsActions = (
   const userKey = getUserControllerMeQueryKey();
 
   const { mutate, isLoading: isMutating } =
-    useInvitationControllerUpdateInvitation({
-      onSettled: () => {
-        queryClient.invalidateQueries(invitationsKey);
-        queryClient.invalidateQueries(userKey);
-      },
-    });
+    useInvitationControllerUpdateInvitation();
 
   const acceptInvite = (id: InvitationDTO['id']) =>
     mutate(
@@ -65,6 +60,8 @@ export const useReceivedInvitationsActions = (
             t('organization.invitations.notifications.rejectedSuccess'),
             NotificationTypeEnum.Success
           );
+          queryClient.invalidateQueries(invitationsKey);
+          queryClient.invalidateQueries(userKey);
         },
         onError: (error: any) => {
           showNotification(
