+++ packages/ui/libs/organization/data/src/handlers/registerIRec.ts
@@ -1,5 +1,4 @@
 import {
-  getRegistrationControllerGetRegistrationsQueryKey,
   IRECAccountType,
   NewRegistrationDTO,
   useRegistrationControllerRegister,
@@ -10,8 +9,6 @@ import {
   showNotification,
 } from '@energyweb/origin-ui-core';
 import { useTranslation } from 'react-i18next';
-import { useQueryClient } from 'react-query';
-import { useNavigate } from 'react-router-dom';
 
 export type IRecRegistrationInfoForm = {
   accountType: IRECAccountType;
@@ -56,10 +53,6 @@ export type IRecRegisterFormMergedType = IRecRegistrationInfoForm &
 export const useIRecRegisterHandler = (openRegisteredModal: () => void) => {
   const { mutate } = useRegistrationControllerRegister();
   const { t } = useTranslation();
-  const navigate = useNavigate();
-
-  const queryClient = useQueryClient();
-  const iRecOrgKey = getRegistrationControllerGetRegistrationsQueryKey();
 
   return (values: IRecRegisterFormMergedType) => {
     const formattedData: NewRegistrationDTO = {
@@ -67,12 +60,13 @@ export const useIRecRegisterHandler = (openRegisteredModal: () => void) => {
       headquarterCountry: values.headquarterCountry
         .map((option) => option.value)[0]
         .toString(),
-      primaryContactOrganizationCountry: values.primaryContactOrganizationCountry
-        .map((option) => option.value)[0]
-        .toString(),
+      primaryContactOrganizationCountry:
+        values.primaryContactOrganizationCountry
+          .map((option) => option.value)[0]
+          .toString(),
       // @should be changed on backend and migrated
       // from int to string because of auto-gen client/enum issues
-      accountType: (values.accountType as unknown) as IRECAccountType,
+      accountType: values.accountType as unknown as IRECAccountType,
       registrationYear: Number(values.registrationYear),
       activeCountries: values.activeCountries.map((i) => i?.value as string),
       leadUserTitle:
@@ -83,11 +77,7 @@ export const useIRecRegisterHandler = (openRegisteredModal: () => void) => {
     mutate(
       { data: formattedData },
       {
-        onSuccess: () => {
-          navigate('/organization/my');
-          openRegisteredModal();
-          queryClient.invalidateQueries(iRecOrgKey);
-        },
+        onSuccess: () => openRegisteredModal(),
         onError: (error: any) => {
           showNotification(
             `${t('organization.registerIRec.notifications.registerError')}:
