+++ packages/ui/libs/organization/view/src/containers/modals/IRecRegisteredThankYou/IRecRegisteredThankYou.effects.ts
@@ -1,5 +1,7 @@
+import { getRegistrationControllerGetRegistrationsQueryKey } from '@energyweb/origin-organization-irec-api-react-query-client';
 import { GenericModalProps } from '@energyweb/origin-ui-core';
 import { useIRecRegisteredThankYouLogic } from '@energyweb/origin-ui-organization-logic';
+import { useQueryClient } from 'react-query';
 import { useNavigate } from 'react-router';
 import {
   OrganizationModalsActionsEnum,
@@ -11,12 +13,15 @@ export const useIRecRegisteredThankYouEffects = () => {
   const { iRecRegisteredThankYou: open } = useOrgModalsStore();
   const dispatchModals = useOrgModalsDispatch();
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
+  const iRecOrgKey = getRegistrationControllerGetRegistrationsQueryKey();
 
   const closeModal = () => {
     dispatchModals({
       type: OrganizationModalsActionsEnum.SHOW_IREC_REGISTERED_THANK_YOU,
       payload: false,
     });
+    queryClient.invalidateQueries(iRecOrgKey);
     navigate('/organization/my');
   };
