+++ packages/ui/libs/organization/view/src/containers/modals/RegisterThankYou/RegisterThankYou.effects.ts
@@ -1,5 +1,7 @@
+import { getUserControllerMeQueryKey } from '@energyweb/origin-backend-react-query-client';
 import { GenericModalProps } from '@energyweb/origin-ui-core';
 import { useRegisterThankYouLogic } from '@energyweb/origin-ui-organization-logic';
+import { useQueryClient } from 'react-query';
 import { useNavigate } from 'react-router';
 import {
   OrganizationModalsActionsEnum,
@@ -11,6 +13,8 @@ export const useRegisterThankYouEffects = () => {
   const { registerThankYou: open } = useOrgModalsStore();
   const dispatchModals = useOrgModalsDispatch();
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
+  const userKey = getUserControllerMeQueryKey();
 
   const closeModal = () => {
     dispatchModals({
@@ -18,6 +22,7 @@ export const useRegisterThankYouEffects = () => {
       payload: false,
     });
     navigate('/organization/my');
+    queryClient.invalidateQueries(userKey);
   };
 
   const { title, text, buttons } = useRegisterThankYouLogic(closeModal);
